Remove global canvas, refactor draw functions to use named params
